requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.3
azure-storage-blob==12.17.0
bcrypt==4.0.1
bitsandbytes==0.41.1
black==23.7.0
boto3==1.28.36
botocore==1.31.36
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
cryptography==41.0.3
datasets==2.13.0
dill==0.3.6
faiss-cpu==1.7.4
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
huggingface-hub==0.16.4
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lit==16.0.6
loguru==0.7.0
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.0rc1
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
pathspec==0.11.2
peft==0.4.0
Pillow==9.3.0
platformdirs==3.10.0
protobuf==4.24.2
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sympy==1.11.1
threadpoolctl==3.2.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
trl==0.4.7
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
xxhash==3.3.0
yarl==1.9.2
